+++ CHANGES.md
@@ -4,22 +4,32 @@
 
 * The core logic isn't in the project skeleton anymore. This allow you to 
 update Clay whitout having to recreate your projects.
+
 * Bugfix: Non-utf8 content now is handled correctly.
+
 * Bugfix: Fix pyton package
 
 
 ## 0.4
 
-* Added a per-project settings file.
+* Added a per-project settings file (settings.json)
 
 
 ## 0.5
 
-* Added a list of all generated views (by default in `_index.html`, but it can
-be changed im the settings.
+* Added a list of all generated views at `_index.html`.
 
 
 ## 0.6
 
-* First static file processors! So far:
-    - cleverCSS (*.ccss). Requires clevercss python library.
+* Static file processors:
+    - less (requires node and less).
+    - scss (Best known as Sass 3).
+    - cleverCSS (*.ccss).
+    - coffeescript (requires node and less).
+
+* A default "not found" page.
+
+* A really nice style for the "_index.html" page.
+
+